Add files via upload
Removed view functions and replaced them with Paramiko function. This removed over 700 lines of code and makes outputs easier to get. The is no need to create a view function for every configuration option. NETCONF does have a similar way to fetch configs but for me replys only with an empty but successful RPC response <data><data>